requirements-windows.txt
requests==2.28.1
beautifulsoup4==4.11.1
twilio==7.16.1
python-dotenv==0.21.0
Flask==2.3.3
APScheduler==3.10.4
pytz==2023.3
# Alternative to psycopg2 for Windows - use SQLite instead
# psycopg2-binary==2.9.7 